refactor:fetchHistoryData.js,submitData.jsに処理を分割した
