+++ src/assets/JS/main.js
@@ -1,12 +1,9 @@
-// Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
-// import { getAnalytics } from "firebase/analytics";
-import { getFirestore, collection, getDocs, addDoc } from 'firebase/firestore';
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
+import { getFirestore } from 'firebase/firestore';
+import fetchReports from './modules/fetchHistoryData.js'
+import addReport from './modules/submitData.js'
 
 // Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
 const firebaseConfig = {
   apiKey: import.meta.env.VITE_API_KEY,
   authDomain: import.meta.env.VITE_AUTH_DOMAIN,
@@ -19,71 +16,21 @@ const firebaseConfig = {
 
 // Initialize Firebase
 const app = initializeApp(firebaseConfig);
-console.log(app)
-// const analytics = getAnalytics(app);
+const db = getFirestore(app);
 
 const $history = document.getElementById('js-history');
 
-const db = getFirestore(app);
-console.log(db)
-// Get a list of reports from your database
-const fetchReports = async () => {
-  //Get Collection
-  const reportSnapshot = await getDocs(collection(db, 'reports'));
-  console.log(reportSnapshot)
-  let tags = '';
-  reportSnapshot.forEach(doc => {
-    console.log('id=>', doc.id)
-    tags += `<tr>
-      <td>${doc.data().date}</td>
-      <td>${doc.data().name}</td>
-      <td>${doc.data().work}</td>
-      <td>${doc.data().comment}</td>
-      </tr>`
-  });
-  $history.innerHTML = tags;
-}
-
 if ($history) {
-  fetchReports();
+  fetchReports(db, $history);
 }
-
-
-
-//Create new data
-
-const getFormData = () => ({
-  name: document.getElementById('name').value,
-  date: new Date(),
-  work: document.getElementById('work').value,
-  comment: document.getElementById('comment').value,
-});
-
-
-const addReport = async () => {
-  try {
-    const { name, work, comment, date } = getFormData()
-    const docRef = await addDoc(collection(db, "reports"), {
-      date,
-      name,
-      work,
-      comment
-    });
-    console.log(date,name,work,comment);
-    console.log("Document written with ID: ", docRef.id);
-  } catch (e) {
-    console.error("Error adding document: ", e);
-  }
-} 
-
 
 const $form = document.getElementById('js-form');
 
-if($form){
-  $form .addEventListener('click', (e) => {
+if ($form) {
+  $form.addEventListener('click', (e) => {
     e.preventDefault();
-    addReport()
-    $form .reset();
+    addReport(db)
+    $form.reset();
   });
 }
